requirement.txt

mysql-connector-python==8.3.0


redis==5.0.3

datetime

pandas==2.0.0
numpy<2.0.0

minio==7.2.7
kafka-python-ng==2.2.2

Faker==24.4.0
python-dotenv==1.0.1